fix: delete print & move button to GropuPgae
