+++ src/utils.js
@@ -13,15 +13,12 @@ export const formatDateString = (dateString) => {
 };
 
 export const shareToKakaotalk = () => {
-  const key = import.meta.env.VITE_KAKAO_KEY;
-  // kakao sdk script 를 window.Kakao로 접근
+  const key = import.meta.env.VITE_KAKAO_JS_KEY;
   if (window.Kakao) {
     const kakao = window.Kakao;
 
-    // 중복 initialization 방지
-    // 카카오에서 제공하는 javascript key를 이용하여 initialize
     if (!kakao.isInitialized()) {
-      kakao.init("8cec546a8802b8b706beb1ffb28b0c8a");
+      kakao.init(key);
     }
 
     kakao.Link.sendDefault({
@@ -29,7 +26,7 @@ export const shareToKakaotalk = () => {
       content: {
         title: "PrayU",
         description: "PrayU와 함께 기도해요!",
-        imageUrl: "이미지 주소", // 온보딩 이미지 삽입
+        imageUrl: "이미지 주소",
         link: {
           mobileWebUrl: import.meta.env.VITE_BASE_URL,
           webUrl: import.meta.env.VITE_BASE_URL,
